# FundooNotes/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with PHP Laravel",
        "description": "Integrate Swagger in Laravel application",
        "contact": {
            "email": "rushipatil6632@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/createLabel": {
            "post": {
                "summary": "create label",
                "description": "create label",
                "operationId": "5fa93c835d46a9760f3ccb556db3e8ac",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "labelname"
                                ],
                                "properties": {
                                    "labelname": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Label Added Sucessfully"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/readLabelById": {
            "get": {
                "summary": "Read Label",
                "description": "Read Label For an Particular User",
                "operationId": "c9fa21e19cc8fa4d994fe659d501bb5d",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "All label are found Successfully"
                    },
                    "404": {
                        "description": "Label Not Found"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/readLabels": {
            "get": {
                "summary": "Read All Labels",
                "description": "Read Labels",
                "operationId": "9ea0c1e58d75d6aaaeb4106992d6a628",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Labels Retrieved Successfully."
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    },
                    "404": {
                        "description": "Label not found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/updateLabelById": {
            "post": {
                "summary": "update label",
                "description": "update user label",
                "operationId": "caf842567e4f993a6bdf8c55295361ee",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "labelname"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "labelname": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Label Updated Successfully"
                    },
                    "404": {
                        "description": "Please enter valid id"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/deleteLabel": {
            "post": {
                "summary": "Delete Label",
                "description": "Delete User Label",
                "operationId": "c7ccd12c19dd67f1e7c7f07435be377d",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label Successfully Deleted"
                    },
                    "404": {
                        "description": "Enter valid id"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "laravel bearer authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}